% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.binary.R
\name{as.binary}
\alias{as.binary}
\title{Integer To Binary Vector}
\usage{
as.binary(number, n)
}
\arguments{
\item{number}{Integer. The value to convert to binary.}

\item{n}{Integer. The number of bits to produce.}
}
\value{
Integer vector of length `n` containing 0/1 digits.
}
\description{
Converts a non-negative integer to a length-`n` binary vector using
big-endian ordering (index 1 is the most-significant bit).
}
\examples{
as.binary(5, 4)  # 0101
}
\keyword{internal}
